デザインができました

+++ LearnTrack/app/Http/Controllers/AnswerController.php
@@ -9,17 +9,21 @@ use Illuminate\Support\Facades\Storage;
 use App\Models\Answer;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\File;
-use Illuminate\Support\Facades\Log;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class AnswerController extends Controller
 {
+    use AuthorizesRequests;
     public function create($id) {
         $question = Question::findOrFail($id);
+        $this->authorize('create', [Answer::class, $question]);
         return view('answer.create', compact('question'), ['mode' => 'input']);
     }
     public function store(AnswerRequest $request) {
+        $question = Question::findOrFail($request->question_id);
+        $this->authorize('create', [Answer::class, $question]);
         $input = $request->validated();
-
+        $user = Auth::user();
         if ($request->input('remove_image') === 'true') {
             if (session('confirm_image_path')) {
                 Storage::disk('public')->delete(session('confirm_image_path'));
@@ -41,32 +45,37 @@ class AnswerController extends Controller
         $mode = $request->input('mode');
         $question = Question::findOrFail($input['question_id']);
         if ($mode === 'confirm') {
-            return view('answer.create', ['mode' => 'confirm', 'questionInput' => $question, 'input' => $input]);
+            return view('answer.create', ['mode' => 'confirm', 'questionInput' => $question, 'input' => $input, 'user' => $user]);
         }
         if ($mode === 'edit') {
             return view('answer.create', ['mode' => 'input', 'questionInput' => $question, 'input' => $input]);
         }
         if ($mode === 'post') {
-            $imagePath = null;
+            $imageUrl = null;
             if (session('confirm_image_path')) {
-                $tempPath = session('confirm_image_path');
-                dd($tempPath);
-                $fullTempPath = storage_path('app/public/' . $tempPath);
-                $s3Path = Storage::disk('s3')->putFile('image', new File($fullTempPath));
-                $imagePath = Storage::disk('s3')->url($s3Path);
+                $confirmPath = session('confirm_image_path');
+                $fullConfirmPath = storage_path('app/public/' . $confirmPath);
+                $s3Url = Storage::disk('s3')->putFile('uploads/answer_images', new File($fullConfirmPath));
+                $imageUrl = Storage::disk('s3')->url($s3Url);
 
                 //一時保存のデータを削除
                 Storage::disk('public')->delete(session('confirm_image_path'));
                 session()->forget('confirm_image_path');
             }
-
             Answer::create([
-                'user_id' => Auth::id(),
+                'user_id' => $user->id,
                 'question_id' => $input['question_id'],
                 'content' => $request->content,
-                'image_url' => $imagePath,
+                'image_url' => $imageUrl,
             ]);
         }
         return redirect()->route('question.show',['id' => $input['question_id']]);
+    }
+    public function cancel($id) {
+        if (session('confirm_image_path')) {
+            Storage::disk('public')->delete(session('confirm_image_path'));
+            session()->forget('confirm_image_path');
+        }
+        return redirect()->route('question.show', ['id' => $id]);
     }
 }
